pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Simple app"
authors = [
    {name = "Nikita",email = "nyuskov@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "django-stubs[compatible-mypy] (>=5.2.7,<6.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "pip-system-certs (>=5.3,<6.0)",
    "certifi (>=2025.10.5,<2026.0.0)",
    "django-taggit (>=6.1.0,<7.0.0)",
    "markdown (>=3.9,<4.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "uvicorn-worker (>=0.4.0,<0.5.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^6.0.0"

